// vad-front-app/src/app/repair-ui/models/repair-dashboard.model.ts
export interface RepairDashboard {

  repairName: string;

  location: string;

  totalTickets: number;

  pendingTickets: number;

  acceptedTickets: number;

  completedTickets: number;

  activeRepairs: number;

  completedRepairs: number;

  cancelledRepairs: number;

  totalDemands: number;

  pendingDemands: number;

  acceptedDemands: number;

  rejectedDemands: number;

}